requirements.txt
Django~=5.0
Pillow~=10.0
gunicorn~=22.0
dj-database-url~=2.0
whitenoise~=6.0
djangorestframework~=3.15
djangorestframework-simplejwt~=5.3
